festival calls working using scm file, working on changing voice video working spelling logic working working on revision logic
stats to do
review to do
gui sometimes doesnt show
fix up
